Standardize all analyst agent signal models for consistency
- Fix confidence type from int to float in charlie_munger, nassim_taleb,
  warren_buffett, and news_sentiment agents
- Add Pydantic signal models (BaseModel) to 5 rule-based agents that
  previously returned plain dicts: fundamentals, growth_agent, sentiment,
  technicals, and valuation
- Rename news_sentiment model class from Sentiment to NewsSentimentSignal
  to follow naming convention
- Fix technicals agent message return bug (was duplicating messages via
  state["messages"] + [message] instead of [message])
- Remove unused Field imports from nassim_taleb, warren_buffett, and
  news_sentiment

+++ src/agents/charlie_munger.py
@@ -11,7 +11,7 @@ from src.utils.api_key import get_api_key_from_state
 
 class CharlieMungerSignal(BaseModel):
     signal: Literal["bullish", "bearish", "neutral"]
-    confidence: int
+    confidence: float
     reasoning: str
